refactor: extract mock wallets into dedicated test_utils/mock_wallet module
Instead of gating each item with #[cfg(feature = "manager")] in
wallet.rs, move MockWallet and NonMatchingMockWallet into a separate
mock_wallet.rs module that is conditionally compiled as a whole.

wallet.rs now only contains ManagedWalletInfo::dummy and
TestWalletContext, which have no manager dependency.
