+++ key-wallet/src/test_utils/mod.rs
@@ -1,9 +1,11 @@
 mod account;
+#[cfg(feature = "manager")]
+mod mock_wallet;
 mod utxo;
 mod wallet;
 
 #[cfg(feature = "manager")]
-pub use wallet::MockWallet;
+pub use mock_wallet::MockWallet;
 #[cfg(feature = "manager")]
-pub use wallet::NonMatchingMockWallet;
+pub use mock_wallet::NonMatchingMockWallet;
 pub use wallet::TestWalletContext;
